+++ demisto_sdk/__main__.py
@@ -3391,13 +3391,6 @@ def create_content_graph(
     help="Path to content graph zip file to import",
 )
 @click.option(
-    "-uli",
-    "--use-local-import",
-    is_flag=True,
-    help="Whether to use the current import files to import the graph.",
-    default=False,
-)
-@click.option(
     "-p",
     "--packs",
     help="A comma-separated list of packs to update",
@@ -3424,7 +3417,6 @@ def update_content_graph(
     ctx,
     use_git: bool = False,
     marketplace: MarketplaceVersions = MarketplaceVersions.XSOAR,
-    use_local_import: bool = False,
     imported_path: Path = None,
     packs: list = None,
     no_dependencies: bool = False,
@@ -3436,7 +3428,6 @@ def update_content_graph(
         ctx,
         use_git=use_git,
         marketplace=marketplace,
-        use_local_import=use_local_import,
         imported_path=imported_path,
         packs_to_update=packs,
         no_dependencies=no_dependencies,