+++ demisto_sdk/commands/content_graph/interface/neo4j/queries/import_export.py
@@ -3,6 +3,7 @@ from typing import List
 
 from neo4j import Transaction
 
+from demisto_sdk.commands.content_graph.common import ContentType
 from demisto_sdk.commands.content_graph.interface.neo4j.queries.common import run_query
 
 
@@ -32,12 +33,12 @@ RETURN node""",
 def merge_duplicate_content_items(tx: Transaction) -> None:
     run_query(
         tx,
-        """// Merges possible duplicate content item nodes after import
-MATCH (n:BaseNode{not_in_repository: true})
-MATCH (m:BaseNode{content_type: n.content_type})
+        f"""// Merges possible duplicate content item nodes after import
+MATCH (n:{ContentType.BASE_NODE}{{not_in_repository: true}})
+MATCH (m:{ContentType.BASE_NODE}{{content_type: n.content_type}})
 WHERE ((m.object_id = n.object_id AND m.object_id <> "") OR (m.name = n.name AND m.name <> ""))
 AND m.not_in_repository = false
 WITH m, n
-CALL apoc.refactor.mergeNodes([m, n], {properties: "discard", mergeRels: true}) YIELD node
+CALL apoc.refactor.mergeNodes([m, n], {{properties: "discard", mergeRels: true}}) YIELD node
 RETURN node""",
     )