+++ demisto_sdk/commands/content_graph/parsers/xsiam_dashboard.py
@@ -29,8 +29,6 @@ class XSIAMDashboardParser(
                 "object_id": [
                     "dashboards_data[0].global_id",
                     "dashboards_data[0].id",
-                    "id",
-                    "global_id",
                 ],
                 "fromversion": "fromVersion",
                 "toversion": "toVersion",