+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 # Changelog
 ## Unreleased
+
+## 1.22.0
 * Added Docker Hook support to **pre-commit**; for details see https://github.com/demisto/demisto-sdk/blob/master/demisto_sdk/commands/pre_commit/README.md#docker-hooks
+* Removed **-use-local-import** flag to **graph update** command.
+* Perfomance improvements to **graph** commands.
 * Adjust graph structure to accommodate anticipated changes in infrastructure for the **validate** command.
 * Fixed an issue where the **lint** command with docker, would not give unique container names to different image runs.
 * Started adaption of the validator refactor - for more information, please refer to the [readme files](demisto_sdk/commands/validate/README.md).
+* Added a new `display_name` field to `Pack` entity in graph.
 
 ## 1.21.0
 * Added the argument `--commited-only` to **pre-commit** to skip collecting on staged files.